Move ActionDispatch::IntegrationTestCase into its own file
