+++ actionpack/lib/action_dispatch.rb
@@ -91,7 +91,7 @@ module ActionDispatch
   autoload_under "testing" do
     autoload :Assertions
     autoload :Integration
-    autoload :IntegrationTestCase, "action_dispatch/testing/integration"
+    autoload :IntegrationTestCase, "action_dispatch/integration_test_case"
     autoload :TestProcess
     autoload :TestRequest
     autoload :TestResponse
